+++ OUDS/Core/Tokens/RawTokens/Tests/ElevationRawTokensTests.swift
@@ -29,35 +29,35 @@ final class ElevationRawTokensTests: XCTestCase {
 
     func testRadiusComputation() throws {
         // Given
-        var token = ElevationCompositeRawToken(x: 0, y: 0, blur: 0, color: ColorRawTokens.colorTransparentBlack500)
+        var token = ElevationCompositeRawToken(x: 0, y: 0, blur: 0, color: ColorRawTokens.colorOpacityBlack200)
         // When
         var radius = token.radius
         // Then
         XCTAssertTrue(radius == 0)
 
         // Given
-        token = ElevationCompositeRawToken(x: 0, y: 2, blur: 1, color: ColorRawTokens.colorTransparentBlack400)
+        token = ElevationCompositeRawToken(x: 0, y: 2, blur: 1, color: ColorRawTokens.colorOpacityBlack400)
         // When
         radius = token.radius
         // Then
         XCTAssertTrue(radius == 0.5)
 
         // Given
-        token = ElevationCompositeRawToken(x: 0, y: 3, blur: 2, color: ColorRawTokens.colorTransparentBlack300)
+        token = ElevationCompositeRawToken(x: 0, y: 3, blur: 2, color: ColorRawTokens.colorOpacityBlack160)
         // When
         radius = token.radius
         // Then
         XCTAssertTrue(radius == 1)
 
         // Given
-        token = ElevationCompositeRawToken(x: 0, y: 4, blur: 4, color: ColorRawTokens.colorTransparentBlack500)
+        token = ElevationCompositeRawToken(x: 0, y: 4, blur: 4, color: ColorRawTokens.colorOpacityBlack200)
         // When
         radius = token.radius
         // Then
         XCTAssertTrue(radius == 2)
 
         // Given
-        token = ElevationCompositeRawToken(x: 0, y: 12, blur: 12, color: ColorRawTokens.colorTransparentBlack300)
+        token = ElevationCompositeRawToken(x: 0, y: 12, blur: 12, color: ColorRawTokens.colorOpacityBlack160)
         // When
         radius = token.radius
         // Then